LiveRegisterPass: attach liveregister.killed to terminator instr
to prepare for LLVM 3.6. From this version on, Metadata does no
longer inherit from Value. With this change it is no longer
possible to use BasicBlock* to label incoming edges (they are not
supported by ValueAsMetadata). Instead, we will use BlockAddress
to determine the edge. But, as it is not allowed to take a
BlockAddress from the entry BasicBlock, we have to annotate based
on outgoing instead of incoming edges.