fix: use users() instead of uses() starting from LLVM 3.5
With version 3.5, LLVM introduced a subtle semantic change in the
API of `Value`. With that change, `use_begin()` is renamed to
`user_begin()`. Additionally, a new method `use_begin()` with a
different meaning was introduced, both with corresponding range
iterators (`uses()´ and `users()`). Now, `use_begin()` actually
iterates over `Use`s, whereas before, dereferencing it would give
a `User *`. See https://reviews.llvm.org/rL203364 for details.

Due to the reintroduction of `use_begin()`, existing code may still
compile, although the semantics have changed. In the code changed
with this patch, the iterator is dereferenced and the result is
casted to an `Instruction *` (if possible). As `Use` in contrast to
`User` is not a superclass of `Instruction`, this does not work,
leading to a situation where the code inside the if is never
executed.